@@ -1,7 +1,9 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Collections.Generic;
+#if !CSLA_NETCORE
 using System.Windows.Data;
+#endif
 using System.Linq;
 using System.Text;
 using Csla;
@@ -11,11 +13,136 @@ using Csla.Serialization;
 using Csla.Silverlight;
 //using Csla.Validation;
 
-#if!SILVERLIGHT
+#if !SILVERLIGHT
 using System.Data.SqlClient;
 using BusinessLibrary.Data;
 using BusinessLibrary.Security;
 
+#endif
+
+#if CSLA_NETCORE
+namespace System.ComponentModel
+{
+    //
+    // Summary:
+    //     Defines methods and properties that a collection view implements to provide paging
+    //     capabilities to a collection.
+    public interface IPagedCollectionView
+    {
+        //
+        // Summary:
+        //     Gets a value that indicates whether the System.ComponentModel.IPagedCollectionView.PageIndex
+        //     value can change.
+        //
+        // Returns:
+        //     true if the System.ComponentModel.IPagedCollectionView.PageIndex value can change;
+        //     otherwise, false.
+        bool CanChangePage { get; }
+        //
+        // Summary:
+        //     Gets a value that indicates whether the page index is changing.
+        //
+        // Returns:
+        //     true if the page index is changing; otherwise, false.
+        bool IsPageChanging { get; }
+        //
+        // Summary:
+        //     Gets the number of known items in the view before paging is applied.
+        //
+        // Returns:
+        //     The number of known items in the view before paging is applied.
+        int ItemCount { get; }
+        //
+        // Summary:
+        //     Gets the zero-based index of the current page.
+        //
+        // Returns:
+        //     The zero-based index of the current page.
+        int PageIndex { get; }
+        //
+        // Summary:
+        //     Gets or sets the number of items to display on a page.
+        //
+        // Returns:
+        //     The number of items to display on a page.
+        int PageSize { get; set; }
+        //
+        // Summary:
+        //     Gets the total number of items in the view before paging is applied.
+        //
+        // Returns:
+        //     The total number of items in the view before paging is applied, or -1 if the
+        //     total number is unknown.
+        int TotalItemCount { get; }
+
+        //
+        // Summary:
+        //     When implementing this interface, raise this event after the System.ComponentModel.IPagedCollectionView.PageIndex
+        //     has changed.
+        event EventHandler<EventArgs> PageChanged;
+        //
+        // Summary:
+        //     When implementing this interface, raise this event before changing the System.ComponentModel.IPagedCollectionView.PageIndex.
+        //     The event handler can cancel this event.
+        event EventHandler<PageChangingEventArgs> PageChanging;
+
+        //
+        // Summary:
+        //     Sets the first page as the current page.
+        //
+        // Returns:
+        //     true if the operation was successful; otherwise, false.
+        bool MoveToFirstPage();
+        //
+        // Summary:
+        //     Sets the last page as the current page.
+        //
+        // Returns:
+        //     true if the operation was successful; otherwise, false.
+        bool MoveToLastPage();
+        //
+        // Summary:
+        //     Moves to the page after the current page.
+        //
+        // Returns:
+        //     true if the operation was successful; otherwise, false.
+        bool MoveToNextPage();
+        //
+        // Summary:
+        //     Moves to the page at the specified index.
+        //
+        // Parameters:
+        //   pageIndex:
+        //     The index of the page to move to.
+        //
+        // Returns:
+        //     true if the operation was successful; otherwise, false.
+        bool MoveToPage(int pageIndex);
+        //
+        // Summary:
+        //     Moves to the page before the current page.
+        //
+        // Returns:
+        //     true if the operation was successful; otherwise, false.
+        bool MoveToPreviousPage();
+    }
+    //
+    // Summary:
+    //     Provides data for the System.ComponentModel.IPagedCollectionView.PageChanging
+    //     event.
+    public sealed class PageChangingEventArgs : CancelEventArgs
+    {
+        private int _page = 1;
+        public PageChangingEventArgs(int newPageIndex)
+        {
+            _page = newPageIndex;
+        }
+        public int NewPageIndex
+        {
+            get { return _page; }
+        }
+    }
+}
 #endif
 
 namespace BusinessLibrary.BusinessClasses
@@ -24,7 +151,7 @@ namespace BusinessLibrary.BusinessClasses
     public class ItemList : DynamicBindingListBase<Item>, System.ComponentModel.IPagedCollectionView, INotifyPropertyChanged
     {
 
-        #region Implementing IPagedCollectionView
+#region Implementing IPagedCollectionView
         public event PropertyChangedEventHandler PropertyChanged;
         private void NotifyPropertyChanged(String info)
         {
@@ -85,7 +212,7 @@ namespace BusinessLibrary.BusinessClasses
         public event EventHandler<EventArgs> PageChanged;
 
         public event EventHandler<System.ComponentModel.PageChangingEventArgs> PageChanging;
-
+        [Newtonsoft.Json.JsonProperty]
         public int PageCount
         {
             get
@@ -100,6 +227,7 @@ namespace BusinessLibrary.BusinessClasses
         }
 
         private int _totalItemCount;
+        [Newtonsoft.Json.JsonProperty]
         public int TotalItemCount
         {
             get
@@ -115,7 +243,9 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
+
         private int _pageSize ;//= 700;
+        [Newtonsoft.Json.JsonProperty]
         public int PageSize
         {
             get
@@ -133,6 +263,7 @@ namespace BusinessLibrary.BusinessClasses
         }
 
         private int _pageIndex;
+        [Newtonsoft.Json.JsonProperty]
         public int PageIndex
         {
             get
@@ -140,7 +271,7 @@ namespace BusinessLibrary.BusinessClasses
                 return _pageIndex;
             }
         }
-
+        [Newtonsoft.Json.JsonProperty]
         public int ItemCount
         {
             get
@@ -167,7 +298,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        #endregion
+#endregion
 
         public bool HasSavedHandler = false;
 
@@ -247,7 +378,7 @@ namespace BusinessLibrary.BusinessClasses
     }
 
 #else
-        private ItemList() { }
+        public ItemList() { }
 #endif
 
 
@@ -477,7 +608,8 @@ namespace BusinessLibrary.BusinessClasses
     [Serializable]
     public class SelectionCriteria : BusinessBase //Csla.CriteriaBase
     {
-        private static PropertyInfo<bool> OnlyIncludedProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("OnlyIncluded", "OnlyIncluded", true));
+        public SelectionCriteria() { }
+        public static readonly PropertyInfo<bool> OnlyIncludedProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("OnlyIncluded", "OnlyIncluded", true));
         public bool OnlyIncluded
         {
             get { return ReadProperty(OnlyIncludedProperty); }
@@ -487,7 +619,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<bool> ShowDeletedProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowDeleted", "ShowDeleted", false));
+        public static readonly PropertyInfo<bool> ShowDeletedProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowDeleted", "ShowDeleted", false));
         public bool ShowDeleted
         {
             get { return ReadProperty(ShowDeletedProperty); }
@@ -497,7 +629,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> SourceIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SourceId", "SourceId", 0));
+        public static readonly PropertyInfo<int> SourceIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SourceId", "SourceId", 0));
         public int SourceId
         {
             get { return ReadProperty(SourceIdProperty); }
@@ -507,7 +639,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> SearchIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SearchId", "SearchId", 0));
+        public static readonly PropertyInfo<int> SearchIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SearchId", "SearchId", 0));
         public int SearchId
         {
             get { return ReadProperty(SearchIdProperty); }
@@ -517,7 +649,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> XAxisSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("XAxisSetId", "XAxisSetId"));
+        public static readonly PropertyInfo<Int64> XAxisSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("XAxisSetId", "XAxisSetId"));
         public Int64 XAxisSetId
         {
             get { return ReadProperty(XAxisSetIdProperty); }
@@ -527,7 +659,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> XAxisAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("XAxisAttributeId", "XAxisAttributeId"));
+        public static readonly PropertyInfo<Int64> XAxisAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("XAxisAttributeId", "XAxisAttributeId"));
         public Int64 XAxisAttributeId
         {
             get { return ReadProperty(XAxisAttributeIdProperty); }
@@ -537,7 +669,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> YAxisSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("YAxisSetId", "YAxisSetId"));
+        public static readonly PropertyInfo<Int64> YAxisSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("YAxisSetId", "YAxisSetId"));
         public Int64 YAxisSetId
         {
             get { return ReadProperty(YAxisSetIdProperty); }
@@ -547,7 +679,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> YAxisAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("YAxisAttributeId", "YAxisAttributeId"));
+        public static readonly PropertyInfo<Int64> YAxisAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("YAxisAttributeId", "YAxisAttributeId"));
         public Int64 YAxisAttributeId
         {
             get { return ReadProperty(YAxisAttributeIdProperty); }
@@ -557,7 +689,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> FilterSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("FilterSetId", "FilterSetId"));
+        public static readonly PropertyInfo<Int64> FilterSetIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("FilterSetId", "FilterSetId"));
         public Int64 FilterSetId
         {
             get { return ReadProperty(FilterSetIdProperty); }
@@ -567,7 +699,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<Int64> FilterAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("FilterAttributeId", "FilterAttributeId"));
+        public static readonly PropertyInfo<Int64> FilterAttributeIdProperty = RegisterProperty<Int64>(typeof(SelectionCriteria), new PropertyInfo<Int64>("FilterAttributeId", "FilterAttributeId"));
         public Int64 FilterAttributeId
         {
             get { return ReadProperty(FilterAttributeIdProperty); }
@@ -577,7 +709,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<string> AttributeSetIdListProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("AttributeSetIdList", "AttributeSetIdList", string.Empty));
+        public static readonly PropertyInfo<string> AttributeSetIdListProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("AttributeSetIdList", "AttributeSetIdList", string.Empty));
         public string AttributeSetIdList
         {
             get { return ReadProperty(AttributeSetIdListProperty); }
@@ -587,7 +719,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<string> ListTypeProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("ListType", "ListType", "StandardItemList"));
+        public static readonly PropertyInfo<string> ListTypeProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("ListType", "ListType", "StandardItemList"));
         public string ListType
         {
             get { return ReadProperty(ListTypeProperty); }
@@ -597,7 +729,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> PageNumberProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("PageNumber", "PageNumber", 0));
+        public static readonly PropertyInfo<int> PageNumberProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("PageNumber", "PageNumber", 0));
         public int PageNumber
         {
             get { return ReadProperty(PageNumberProperty); }
@@ -607,7 +739,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> PageSizeProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("PageSize", "PageSize", 1));
+        public static readonly PropertyInfo<int> PageSizeProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("PageSize", "PageSize", 1));
         public int PageSize
         {
             get { return ReadProperty(PageSizeProperty); }
@@ -617,7 +749,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> WorkAllocationIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("WorkAllocationId", "WorkAllocationId", 0));
+        public static readonly PropertyInfo<int> WorkAllocationIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("WorkAllocationId", "WorkAllocationId", 0));
         public int WorkAllocationId
         {
             get { return ReadProperty(WorkAllocationIdProperty); }
@@ -627,7 +759,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> ComparisonIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("ComparisonId", "ComparisonId", 0));
+        public static readonly PropertyInfo<int> ComparisonIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("ComparisonId", "ComparisonId", 0));
         public int ComparisonId
         {
             get { return ReadProperty(ComparisonIdProperty); }
@@ -637,7 +769,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<string> DescriptionProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("Description", "Description", string.Empty));
+        public static readonly PropertyInfo<string> DescriptionProperty = RegisterProperty<string>(typeof(SelectionCriteria), new PropertyInfo<string>("Description", "Description", string.Empty));
         public string Description
         {
             get { return ReadProperty(DescriptionProperty); }
@@ -647,7 +779,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> ContactIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("ContactId", "ContactId", 0));
+        public static readonly PropertyInfo<int> ContactIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("ContactId", "ContactId", 0));
         public int ContactId
         {
             get { return ReadProperty(ContactIdProperty); }
@@ -657,7 +789,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<int> SetIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SetId", "SetId", 0));
+        public static readonly PropertyInfo<int> SetIdProperty = RegisterProperty<int>(typeof(SelectionCriteria), new PropertyInfo<int>("SetId", "SetId", 0));
         public int SetId
         {
             get { return ReadProperty(SetIdProperty); }
@@ -667,7 +799,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<bool> ShowInfoColumnProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowInfoColumn", "ShowInfoColumn", false));
+        public static readonly PropertyInfo<bool> ShowInfoColumnProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowInfoColumn", "ShowInfoColumn", false));
         public bool ShowInfoColumn
         {
             get { return ReadProperty(ShowInfoColumnProperty); }
@@ -677,7 +809,7 @@ namespace BusinessLibrary.BusinessClasses
             }
         }
 
-        private static PropertyInfo<bool> ShowScoreColumnProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowScoreColumn", "ShowScoreColumn", false));
+        public static readonly PropertyInfo<bool> ShowScoreColumnProperty = RegisterProperty<bool>(typeof(SelectionCriteria), new PropertyInfo<bool>("ShowScoreColumn", "ShowScoreColumn", false));
         public bool ShowScoreColumn
         {
             get { return ReadProperty(ShowScoreColumnProperty); }
@@ -686,8 +818,5 @@ namespace BusinessLibrary.BusinessClasses
                 SetProperty(ShowScoreColumnProperty, value);
             }
         }
-
-        public SelectionCriteria() { }
-
     }
 }
