@@ -19,7 +19,7 @@ namespace BusinessLibrary.Security
     [Serializable()]
     public class ReviewerIdentity : CslaIdentity
     {
-        private static readonly PropertyInfo<int> UserIdProperty = RegisterProperty<int>(typeof(ReviewerIdentity), new PropertyInfo<int>("UserId", "User Id", 0));
+        public static readonly PropertyInfo<int> UserIdProperty = RegisterProperty<int>(typeof(ReviewerIdentity), new PropertyInfo<int>("UserId", "User Id", 0));
         public int UserId
         {
             get
@@ -27,7 +27,7 @@ namespace BusinessLibrary.Security
                 return GetProperty<int>(UserIdProperty);
             }
         }
-        private static readonly PropertyInfo<bool> IsSiteAdminProperty = RegisterProperty<bool>(typeof(ReviewerIdentity), new PropertyInfo<bool>("IsSiteAdmin", "IsSiteAdmin", false));
+        public static readonly PropertyInfo<bool> IsSiteAdminProperty = RegisterProperty<bool>(typeof(ReviewerIdentity), new PropertyInfo<bool>("IsSiteAdmin", "IsSiteAdmin", false));
         public bool IsSiteAdmin
         {
             get
@@ -35,7 +35,7 @@ namespace BusinessLibrary.Security
                 return GetProperty<bool>(IsSiteAdminProperty);
             }
         }
-        private static readonly PropertyInfo<int> ReviewIdProperty = RegisterProperty<int>(typeof(ReviewerIdentity), new PropertyInfo<int>("ReviewId", "Review Id", 0));
+        public static readonly PropertyInfo<int> ReviewIdProperty = RegisterProperty<int>(typeof(ReviewerIdentity), new PropertyInfo<int>("ReviewId", "Review Id", 0));
         public int ReviewId
         {
 #if SILVERLIGHT
@@ -85,7 +85,7 @@ namespace BusinessLibrary.Security
 #endif
 
         }
-        private static readonly PropertyInfo<string> TicketProperty = RegisterProperty<string>(typeof(ReviewerIdentity), new PropertyInfo<string>("Ticket", "Ticket", 0));
+        public static readonly PropertyInfo<string> TicketProperty = RegisterProperty<string>(typeof(ReviewerIdentity), new PropertyInfo<string>("Ticket", "Ticket", 0));
         public string Ticket
         {
             get
@@ -93,7 +93,7 @@ namespace BusinessLibrary.Security
                 return GetProperty<string>(TicketProperty);
             }
         }
-        private static readonly PropertyInfo<DateTime> AccountExpirationProperty = RegisterProperty<DateTime>(typeof(ReviewerIdentity), new PropertyInfo<DateTime>("AccountExpiration", "AccountExpiration"));
+        public static readonly PropertyInfo<DateTime> AccountExpirationProperty = RegisterProperty<DateTime>(typeof(ReviewerIdentity), new PropertyInfo<DateTime>("AccountExpiration", "AccountExpiration"));
         public DateTime AccountExpiration
         {
             get
@@ -101,7 +101,7 @@ namespace BusinessLibrary.Security
                 return GetProperty<DateTime>(AccountExpirationProperty);
             }
         }
-        private static readonly PropertyInfo<DateTime> ReviewExpirationProperty = RegisterProperty<DateTime>(typeof(ReviewerIdentity), new PropertyInfo<DateTime>("ReviewExpiration", "ReviewExpiration"));
+        public static readonly PropertyInfo<DateTime> ReviewExpirationProperty = RegisterProperty<DateTime>(typeof(ReviewerIdentity), new PropertyInfo<DateTime>("ReviewExpiration", "ReviewExpiration"));
         public DateTime ReviewExpiration
         {
             get
@@ -109,7 +109,7 @@ namespace BusinessLibrary.Security
                 return GetProperty<DateTime>(ReviewExpirationProperty);
             }
         }
-        private static readonly PropertyInfo<string> LoginModeProperty = RegisterProperty<string>(typeof(ReviewerIdentity), new PropertyInfo<string>("LoginMode", "LoginMode", 0));
+        public static readonly PropertyInfo<string> LoginModeProperty = RegisterProperty<string>(typeof(ReviewerIdentity), new PropertyInfo<string>("LoginMode", "LoginMode", 0));
         public string LoginMode
         {
             get
@@ -136,6 +136,21 @@ namespace BusinessLibrary.Security
                 return (IsInRole("CochraneUser"));
             }
         }
+        public int DaysLeftAccount
+        {
+            get
+            {
+                return AccountExpiration.Subtract(DateTime.Today).Days;
+            }
+        }
+        public int DaysLeftReview
+        {
+            get
+            {
+                return ReviewExpiration.Subtract(DateTime.Today).Days;
+            }
+        }
+
 #if SILVERLIGHT
 
     public ReviewerIdentity() {}
@@ -152,25 +167,34 @@ namespace BusinessLibrary.Security
     {
         GetCslaIdentity<ReviewerIdentity>(completed, new CredentialsCriteria(username, password, ArchieCode, Status, LoginMode, reviewId));
     }
-    public int DaysLeftAccount
-    {
-        get
-        {
-            return AccountExpiration.Subtract(DateTime.Today).Days;
-        }
-    }
-    public int DaysLeftReview
-    {
-        get
-        {
-            return ReviewExpiration.Subtract(DateTime.Today).Days;
-        }
-    }
 #else
-        public static void GetIdentity(string username, string password, int reviewId, string LoginMode, string roles)
+
+        public static ReviewerIdentity GetIdentity(string username, string password, int reviewId, string LoginMode, string roles)
         {
-            GetCslaIdentity<ReviewerIdentity>(new CredentialsCriteria(username, password, reviewId, LoginMode));
+            return GetCslaIdentity<ReviewerIdentity>(new CredentialsCriteria(username, password, reviewId, LoginMode));
         }
+#if (CSLA_NETCORE)
+        public static ReviewerIdentity GetIdentity(int contactId, int reviewId, string displayName)
+        {
+            return GetCslaIdentity<ReviewerIdentity>(new CredentialsCriteria(contactId, reviewId, displayName, "CSLA"));
+        }
+        public static ReviewerIdentity GetIdentity(System.Security.Claims.ClaimsPrincipal CP)
+        {
+            return GetCslaIdentity<ReviewerIdentity>(new CredentialsCriteria(CP));
+        }
+        public static readonly PropertyInfo<string> TokenProperty = RegisterProperty<string>(typeof(ReviewerIdentity), new PropertyInfo<string>("Token", "Token", 0));
+        public string Token
+        {
+            get
+            {
+                return GetProperty<string>(TokenProperty);
+            }
+            set
+            {
+                LoadProperty(TokenProperty, value);
+            }
+        }
+#endif
         DateTime ContactExp = new DateTime(1, 1, 1);
         DateTime ReviewExp = new DateTime(1, 1, 1);
         private void DataPortal_Fetch(CredentialsCriteria criteria)
@@ -262,69 +286,7 @@ namespace BusinessLibrary.Security
                                         //a. user is not Cochrane (no Archie IDs in ER4)
                                         //b. user has an Archie Id but tokens couldn't be refreshed
                                         //c. user was in Archie but isn't anymore.
-
-                                        using (SqlCommand command2 = new SqlCommand("st_ContactLoginReview", connection))
-                                        {
-                                            DateTime check;
-                                            command2.CommandType = System.Data.CommandType.StoredProcedure;
-                                            command2.Parameters.Add(new SqlParameter("@userId", UserId));
-                                            command2.Parameters.Add(new SqlParameter("@reviewId", criteria.ReviewId));
-                                            command2.Parameters.Add(new SqlParameter("@IsArchieUser", false));
-                                            command2.Parameters.Add("@GUI", System.Data.SqlDbType.UniqueIdentifier);
-                                            command2.Parameters["@GUI"].Direction = System.Data.ParameterDirection.Output;
-                                            command2.CommandTimeout = 60;
-                                            using (Csla.Data.SafeDataReader reader2 = new Csla.Data.SafeDataReader(command2.ExecuteReader()))
-                                            {
-                                                if (reader2.Read())
-                                                {
-                                                    if (criteria.ReviewId == reader2.GetInt32("REVIEW_ID")
-                                                       && (reader2.GetString("ARCHIE_ID") == "" || reader2.GetString("ARCHIE_ID") == "prospective_______")
-                                                        //we do not allow non archie users to access registered Archie reviews, 
-                                                        //but we do allow non archie users to open prospective archie reviews
-                                                       )
-                                                    {
-                                                        ContactExp = reader2.GetDateTime("CONTACT_EXP");
-                                                        ReviewExp = reader2.GetDateTime("REVIEW_EXP");
-                                                        if (ReviewExp == new DateTime(1, 1, 1)) ReviewExp = new DateTime(3000, 1, 1);
-                                                        LoadProperty<int>(ReviewIdProperty, criteria.ReviewId);
-                                                        //if (check == new DateTime(1, 1, 1)) check = DateTime.Now.AddDays(1);
-                                                        check = ContactExp < ReviewExp ? ContactExp : ReviewExp;
-                                                        if (check < DateTime.Today && (reader2.GetInt32("FUNDER_ID") != UserId || ContactExp < DateTime.Today))
-                                                        {//if either the account or the review are expired and (this is not the review owner, or the current account is expired) then we 
-                                                            //mark the user as ReadOnly. 
-                                                            //Hence, if the the review is expired, the account is not, and the user is the review owner, the user DOES NOT get the ReadOnly restrictions
-                                                            Roles.Add("ReadOnlyUser");
-                                                        }
-
-                                                        Roles.Add(reader2.GetString("Role"));
-                                                        while (reader2.Read())//query returns multiple lines, one per Role, so in case user has many roles, we read all lines
-                                                        {
-                                                            Roles.Add(reader2.GetString("Role"));
-                                                        }
-                                                        //Roles.Add("AdminUser");// ADDED BY JAMES TEMPORARILY!!!
-
-                                                    }
-                                                    else
-                                                    {
-                                                        LoadProperty<int>(ReviewIdProperty, 0);
-                                                        IsAuthenticated = false;
-                                                    }
-                                                }
-                                                else
-                                                {
-                                                    LoadProperty<int>(ReviewIdProperty, 0);
-                                                    IsAuthenticated = false;
-                                                }
-                                            }
-
-
-                                            if (GetProperty<int>(ReviewIdProperty) != 0)
-                                            {
-                                                LoadProperty(TicketProperty, command2.Parameters["@GUI"].Value.ToString());
-                                                LoadProperty(ReviewExpirationProperty, ReviewExp);
-                                                LoadProperty(AccountExpirationProperty, ContactExp);
-                                            }
-                                        }
+                                        LoginToReview(connection, criteria);
                                     }
                                 }
                                 
@@ -334,7 +296,13 @@ namespace BusinessLibrary.Security
                     connection.Close();
                 }
             }
-            else if (criteria.ArchieState != "" && criteria.ArchieCode != "" && (criteria.Password == null || criteria.Password == ""))
+            //!!!NOTE: this hasn't been proved to work in MVC world.
+            else if (criteria.ArchieState != "" 
+                && criteria.ArchieState != null
+                && criteria.ArchieCode != "" 
+                && criteria.ArchieCode != null
+                && (criteria.Password == null || criteria.Password == "")
+                )
             {//authenticating via Archie
                 if (criteria.ReviewId == 0)
                 {
@@ -501,6 +469,7 @@ namespace BusinessLibrary.Security
 
                 }
             }
+            //!!!NOTE: this hasn't been proved to work in MVC world.
             else if (criteria.Password != null && criteria.Password != "" 
                 && criteria.Username != null && criteria.Username != "" 
                 && criteria.ArchieCode == null || criteria.ArchieCode !=""
@@ -563,9 +532,131 @@ namespace BusinessLibrary.Security
                     LoadProperty<bool>(IsSiteAdminProperty, false);
                 }
             }
-
+#if (CSLA_NETCORE)
+            else if (criteria.ReviewId != 0 && criteria.ContactId != 0)
+            {
+                using (SqlConnection connection = new SqlConnection(DataConnection.ConnectionString))
+                {
+                    connection.Open();
+                    Roles = new MobileList<string>();
+                    LoadProperty<int>(UserIdProperty, criteria.ContactId);
+                    LoginToReview(connection, criteria);
+                    if (ReviewId != 0 && Roles.Count > 0)
+                    {//it worked
+                        LoadProperty<int>(UserIdProperty, criteria.ContactId);
+                        LoadProperty<string>(NameProperty, criteria.DisplayName);
+                        IsAuthenticated = true;
+                    }
+                    else
+                    {//logon to review failed!
+                        LoadProperty<int>(ReviewIdProperty, 0);
+                        LoadProperty<int>(UserIdProperty, 0);
+                        IsAuthenticated = false;
+                    }
+                }
+            }
+            else if (criteria.ClaimsP != null && criteria.ClaimsP.Claims != null && criteria.ClaimsP.Claims.Count() > 0)
+            {//build the RI object quickly based on the data we got from the JWT, WITHOUT passing through the DB.
+                //this is used when an MVC controller will rely on a CSLA BO that needs a fully formed RI object.
+                //as a result, the Ticket will be checked against the DB if and where RI.ReviewId is retreived, as in ER4.
+                //thus, it's OK to avoid checking on the DB at this stage.
+                Roles = new MobileList<string>();
+                foreach (System.Security.Claims.Claim claim in criteria.ClaimsP.Claims)
+                {
+                    switch (claim.Type)
+                    {
+                        case System.Security.Claims.ClaimTypes.Role:
+                            Roles.Add(claim.Value);
+                            break;
+                        case "reviewId":
+                            int tmp;
+                            if (int.TryParse(claim.Value, out tmp)) LoadProperty<int>(ReviewIdProperty, tmp);
+                            break;
+                        case "userId":
+                            int tmp2;
+                            if (int.TryParse(claim.Value, out tmp2)) LoadProperty<int>(UserIdProperty, tmp2);
+                            break;
+                        case "name":
+                            LoadProperty<string>(NameProperty, claim.Value);
+                            break;
+                        case "reviewTicket":
+                            LoadProperty<string>(TicketProperty, claim.Value);
+                            break;
+                        case "isSiteAdmin":
+                            LoadProperty<bool>(IsSiteAdminProperty, (claim.Value.ToLower() == "true"));
+                            break;
+                        default:
+                            break;
+                    }
+                }
+            }
+#endif
         }
+        private void LoginToReview(SqlConnection connection, CredentialsCriteria criteria)
+        {
+            using (SqlCommand command2 = new SqlCommand("st_ContactLoginReview", connection))
+            {
+                DateTime check;
+                command2.CommandType = System.Data.CommandType.StoredProcedure;
+                command2.Parameters.Add(new SqlParameter("@userId", UserId));
+                command2.Parameters.Add(new SqlParameter("@reviewId", criteria.ReviewId));
+                command2.Parameters.Add(new SqlParameter("@IsArchieUser", false));
+                command2.Parameters.Add("@GUI", System.Data.SqlDbType.UniqueIdentifier);
+                command2.Parameters["@GUI"].Direction = System.Data.ParameterDirection.Output;
+                command2.CommandTimeout = 60;
+                using (Csla.Data.SafeDataReader reader2 = new Csla.Data.SafeDataReader(command2.ExecuteReader()))
+                {
+                    if (reader2.Read())
+                    {
+                        if (criteria.ReviewId == reader2.GetInt32("REVIEW_ID")
+                           && (reader2.GetString("ARCHIE_ID") == "" || reader2.GetString("ARCHIE_ID") == "prospective_______")
+                           //we do not allow non archie users to access registered Archie reviews, 
+                           //but we do allow non archie users to open prospective archie reviews
+                           )
+                        {
+                            ContactExp = reader2.GetDateTime("CONTACT_EXP");
+                            ReviewExp = reader2.GetDateTime("REVIEW_EXP");
+                            if (ReviewExp == new DateTime(1, 1, 1)) ReviewExp = new DateTime(3000, 1, 1);
+                            LoadProperty<int>(ReviewIdProperty, criteria.ReviewId);
+                            //if (check == new DateTime(1, 1, 1)) check = DateTime.Now.AddDays(1);
+                            check = ContactExp < ReviewExp ? ContactExp : ReviewExp;
+                            if (check < DateTime.Today && (reader2.GetInt32("FUNDER_ID") != UserId || ContactExp < DateTime.Today))
+                            {//if either the account or the review are expired and (this is not the review owner, or the current account is expired) then we 
+                             //mark the user as ReadOnly. 
+                             //Hence, if the the review is expired, the account is not, and the user is the review owner, the user DOES NOT get the ReadOnly restrictions
+                                Roles.Add("ReadOnlyUser");
+                            }
 
+                            Roles.Add(reader2.GetString("Role"));
+                            while (reader2.Read())//query returns multiple lines, one per Role, so in case user has many roles, we read all lines
+                            {
+                                Roles.Add(reader2.GetString("Role"));
+                            }
+                            //Roles.Add("AdminUser");// ADDED BY JAMES TEMPORARILY!!!
+
+                        }
+                        else
+                        {
+                            LoadProperty<int>(ReviewIdProperty, 0);
+                            IsAuthenticated = false;
+                        }
+                    }
+                    else
+                    {
+                        LoadProperty<int>(ReviewIdProperty, 0);
+                        IsAuthenticated = false;
+                    }
+                }
+
+
+                if (GetProperty<int>(ReviewIdProperty) != 0)
+                {
+                    LoadProperty(TicketProperty, command2.Parameters["@GUI"].Value.ToString());
+                    LoadProperty(ReviewExpirationProperty, ReviewExp);
+                    LoadProperty(AccountExpirationProperty, ContactExp);
+                }
+            }
+        }
         private bool CheckCochraneUser(SqlConnection connection, CredentialsCriteria criteria, ArchieIdentity Ai)
         {
             //with cochrane users, we want:
