@@ -2,9 +2,12 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppModuleShared } from './app.shared.module';
 import { AppComponent } from './components/app/app.component';
-import { ReviewerIdentityService } from './components/app/revieweridentity.service';
+import { ReviewerIdentityService } from './components/services/revieweridentity.service';
+import { JwtInterceptor } from './components/helpers/jwt.interceptor'
 import { HomeComponent } from './components/home/home.component';
-import { FetchReadOnlyReviewsComponent } from './components/readonlyreviews/readonlyreviews.component';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+
+//import { FetchReadOnlyReviewsComponent } from './components/readonlyreviews/readonlyreviews.component';
 
 @NgModule({
     bootstrap: [AppComponent
@@ -13,15 +16,17 @@ import { FetchReadOnlyReviewsComponent } from './components/readonlyreviews/read
     ],
     imports: [
         BrowserModule,
+        HttpClientModule,
         AppModuleShared
     ],
     providers: [
-        ReviewerIdentityService ,
-        { provide: 'BASE_URL', useFactory: getBaseUrl }
+        //ReviewerIdentityService ,
+        { provide: 'BASE_URL', useFactory: getBaseUrl },
+        { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true }
     ]
 })
 export class AppModule {
-    constructor(private ReviewerIdentity: ReviewerIdentityService) { }
+    constructor() { }
 }
 
 export function getBaseUrl() {
