@@ -2,9 +2,34 @@
     ViewData["Title"] = "Home Page";
 }
 
-<app asp-prerender-module="ClientApp/dist/main-server">Loading...</app>
+@*<app asp-prerender-module="ClientApp/dist/main-server">Loading...</app>*@
+<head>
+    <style>
+        .RelevantTerm {
+            color: white;
+            background-color: green;
+        }
 
+        .IrrelevantTerm {
+            color: white;
+            background-color: red;
+        }
+    </style>
+
+</head>
+<app>
+    <p>
+        Loading...
+    </p>
+    <p>Please note that this app <b>cannot work in Internet Explorer</b>.</p>
+    <p>Supported browsers include all fairly recent versions of Edge, FireFox, Safari (including iOS) and Chrome.</p>
+</app>
 <script src="~/dist/vendor.js" asp-append-version="true"></script>
 @section scripts {
+
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+
     <script src="~/dist/main-client.js" asp-append-version="true"></script>
 }
