@@ -1,15 +1,15 @@
 ﻿<!DOCTYPE html>
-<html>
+<html class="h-100">
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] - ERxWebClient</title>
+    <title>EPPI-Reviewer Coding (Beta)</title>
     <base href="~/" />
 
     <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
-    <link rel="stylesheet" href="~/dist/allstyles.css"/>
+    @*<link rel="stylesheet" href="~/dist/allstyles.css"/>*@
 </head>
-<body>
+<body  class="h-100">
     @RenderBody()
 
     @RenderSection("scripts", required: false)
