@@ -1,3 +1,3 @@
 @using ERxWebClient
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Microsoft.AspNetCore.SpaServices
+@*@addTagHelper *, Microsoft.AspNetCore.SpaServices*@
