@@ -1,13 +1,15 @@
 const path = require('path');
 const webpack = require('webpack');
 const merge = require('webpack-merge');
-const AotPlugin = require('@ngtools/webpack').AotPlugin;
+const AngularCompilerPlugin = require('@ngtools/webpack').AngularCompilerPlugin;
 const CheckerPlugin = require('awesome-typescript-loader').CheckerPlugin;
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 
 module.exports = (env) => {
     // Configuration in common to both client-side and server-side bundles
     const isDevBuild = !(env && env.prod);
     const sharedConfig = {
+        mode: 'development',
         stats: { modules: false },
         context: __dirname,
         resolve: { extensions: [ '.js', '.ts' ] },
@@ -43,9 +45,9 @@ module.exports = (env) => {
                 moduleFilenameTemplate: path.relative(clientBundleOutputDir, '[resourcePath]') // Point sourcemap entries to the original file locations on disk
             })
         ] : [
-            // Plugins that apply in production builds only
-            new webpack.optimize.UglifyJsPlugin(),
-            new AotPlugin({
+            // Plugins that apply in production builds only webpack.optimize.UglifyJsPlugin 
+                new UglifyJsPlugin(),
+                new AngularCompilerPlugin({
                 tsConfigPath: './tsconfig.json',
                 entryModule: path.join(__dirname, 'ClientApp/app/app.browser.module#AppModule'),
                 exclude: ['./**/*.server.ts']
@@ -54,31 +56,31 @@ module.exports = (env) => {
     });
 
     // Configuration for server-side (prerendering) bundle suitable for running in Node
-    const serverBundleConfig = merge(sharedConfig, {
-        resolve: { mainFields: ['main'] },
-        entry: { 'main-server': './ClientApp/boot.server.ts' },
-        plugins: [
-            new webpack.DllReferencePlugin({
-                context: __dirname,
-                manifest: require('./ClientApp/dist/vendor-manifest.json'),
-                sourceType: 'commonjs2',
-                name: './vendor'
-            })
-        ].concat(isDevBuild ? [] : [
-            // Plugins that apply in production builds only
-            new AotPlugin({
-                tsConfigPath: './tsconfig.json',
-                entryModule: path.join(__dirname, 'ClientApp/app/app.server.module#AppModule'),
-                exclude: ['./**/*.browser.ts']
-            })
-        ]),
-        output: {
-            libraryTarget: 'commonjs',
-            path: path.join(__dirname, './ClientApp/dist')
-        },
-        target: 'node',
-        devtool: 'inline-source-map'
-    });
+    //const serverBundleConfig = merge(sharedConfig, {
+    //    resolve: { mainFields: ['main'] },
+    //    entry: { 'main-server': './ClientApp/boot.server.ts' },
+    //    plugins: [
+    //        new webpack.DllReferencePlugin({
+    //            context: __dirname,
+    //            manifest: require('./ClientApp/dist/vendor-manifest.json'),
+    //            sourceType: 'commonjs2',
+    //            name: './vendor'
+    //        })
+    //    ].concat(isDevBuild ? [] : [
+    //        // Plugins that apply in production builds only
+    //        new AngularCompilerPlugin({
+    //            tsConfigPath: './tsconfig.json',
+    //            entryModule: path.join(__dirname, 'ClientApp/app/app.server.module#AppModule'),
+    //            exclude: ['./**/*.browser.ts']
+    //        })
+    //    ]),
+    //    output: {
+    //        libraryTarget: 'commonjs',
+    //        path: path.join(__dirname, './ClientApp/dist')
+    //    },
+    //    target: 'node',
+    //    devtool: 'inline-source-map'
+    //});
 
-    return [clientBundleConfig, serverBundleConfig];
+    return [clientBundleConfig];
 };
