@@ -1,5 +1,6 @@
 const path = require('path');
 const webpack = require('webpack');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const merge = require('webpack-merge');
 const treeShakableModules = [
@@ -68,28 +69,28 @@ module.exports = (env) => {
                 name: '[name]_[hash]'
             })
         ].concat(isDevBuild ? [] : [
-            new webpack.optimize.UglifyJsPlugin()
+            new UglifyJsPlugin()
         ])
     });
 
-    const serverBundleConfig = merge(sharedConfig, {
-        target: 'node',
-        resolve: { mainFields: ['main'] },
-        entry: { vendor: allModules.concat(['aspnet-prerendering']) },
-        output: {
-            path: path.join(__dirname, 'ClientApp', 'dist'),
-            libraryTarget: 'commonjs2',
-        },
-        module: {
-            rules: [ { test: /\.css(\?|$)/, use: ['to-string-loader', isDevBuild ? 'css-loader' : 'css-loader?minimize' ] } ]
-        },
-        plugins: [
-            new webpack.DllPlugin({
-                path: path.join(__dirname, 'ClientApp', 'dist', '[name]-manifest.json'),
-                name: '[name]_[hash]'
-            })
-        ]
-    });
+    //const serverBundleConfig = merge(sharedConfig, {
+    //    target: 'node',
+    //    resolve: { mainFields: ['main'] },
+    //    entry: { vendor: allModules.concat(['aspnet-prerendering']) },
+    //    output: {
+    //        path: path.join(__dirname, 'ClientApp', 'dist'),
+    //        libraryTarget: 'commonjs2',
+    //    },
+    //    module: {
+    //        rules: [ { test: /\.css(\?|$)/, use: ['to-string-loader', isDevBuild ? 'css-loader' : 'css-loader?minimize' ] } ]
+    //    },
+    //    plugins: [
+    //        new webpack.DllPlugin({
+    //            path: path.join(__dirname, 'ClientApp', 'dist', '[name]-manifest.json'),
+    //            name: '[name]_[hash]'
+    //        })
+    //    ]
+    //});
 
-    return [clientBundleConfig, serverBundleConfig];
+    return [clientBundleConfig];//, serverBundleConfig];
 }
